fix: register event duplication
